refactor(docs): consolidate home demo with Zero-UI vs React state comparison
The home page previously had two side-by-side demo sections doing the
same job: a single Zero-UI widget ("Try it right here") and a separate
/demo/perf page with a Zero-UI vs React state comparison. The
comparison is the more compelling demo, so promote it to the home page
and drop the standalone /demo/perf route.

While there, add a brand-tinted "renders: N" badge to both panes of
the comparison and the existing hybrid-pattern demo, so the no-rerender
claim is visible without needing React DevTools. The badge uses
suppressHydrationWarning because the count differs intentionally
between the server's single render and the client's first render
(double-invoked under reactStrictMode in dev).

- Move app/(home)/demo/perf/_*.tsx to app/(home)/_components/ and
  rename to PascalCase to match LandingDemo's prior convention.
- Delete LandingDemo and the /demo/perf page.
- Replace LandingDemo with Comparison on the home page; rewrite the
  section copy to describe the new side-by-side; drop the now-stale
  /demo/perf link card from the demo grid.
- Use violet directly (not fd-primary tokens) for the perf badges
  since the perf demo has its own local light/dark theme that's
  independent of the global Fumadocs theme.
- Drop /demo/perf from sitemap.ts.
