+++ examples/demo/app/(home)/demo/real-world/_components.tsx
@@ -134,7 +134,9 @@ function Pane({ title, subtitle, renderCount, children }: { title: string; subti
           <h3 className="text-base font-semibold">{title}</h3>
           <p className="text-fd-muted-foreground mt-1 text-xs">{subtitle}</p>
         </div>
-        <div className="shrink-0 rounded-md bg-fd-accent px-2.5 py-1 font-mono text-xs">
+        <div
+          className="border-fd-primary/30 bg-fd-primary/10 text-fd-primary inline-flex shrink-0 items-center gap-1 rounded-md border px-2.5 py-1 font-mono text-xs"
+          suppressHydrationWarning>
           renders: <span className="font-semibold">{renderCount}</span>
         </div>
       </div>
